using synchronized

+++ multithreading/src/main/java/threads/examples/ThreadConstant.java
@@ -4,7 +4,7 @@ package threads.examples;
  * Created by gbouriga on 22/10/15.
  */
 public class ThreadConstant {
-    public static final int NUMBER_OF_CYCLES = 10;
+    public static final Long NUMBER_OF_CYCLES = 10L;
 
     //no instance
     private ThreadConstant(){
